move files around in cart components folder
