+++ components/cart/order/OrderSection.jsx
@@ -3,7 +3,7 @@ import PropTypes from 'prop-types';
 import { FlatList } from 'react-native';
 
 import OrderItem, { OrderItemPropTypes } from './OrderItem';
-import ItemSeparator from './ItemSeparator';
+import ItemSeparator from '../ItemSeparator';
 
 const OrderSection = ({ items = [] }) => (
     <FlatList
